requirements.txt
gradio
google-generativeai
pillow
requests
PyPDF2
python-dotenv
